button desc, memory & warning fixes
added description to buttons in shop, and started working on memory and warning fixes for UI

+++ header/Components/UIComponents/UIGameHealthComponent.h
@@ -2,5 +2,5 @@
 
 struct UIGameHealthComponent
 {
-	float value = -1.0f;
+	int64_t value = -1.0f;
 };
